feature[global]: Ajout des logs sur action fichiers

+++ BackupLogger/LogEntry.cs
@@ -10,11 +10,11 @@ namespace BackupLogger
         public string TaskName { get; set; }
         public string SourcePath { get; set; }
         public string TargetPath { get; set; }
-        public int FileSize { get; set; }
+        public long FileSize { get; set; }
         public int TransferTime { get; set; }
 
         //constructor
-        public LogEntry(string taskName, string sourcePath, string targetPath, int fileSize, int transferTime)
+        public LogEntry(string taskName, string sourcePath, string targetPath, long fileSize, int transferTime)
         {
             Timestamp = DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss");
             TaskName = taskName;
